requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
amqp==5.1.0
apispec==5.1.1
appdirs==1.4.4
astroid==2.11.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
boto3==1.21.27
botocore==1.24.27
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coloredlogs==15.0.1
cryptography==36.0.2
Deprecated==1.2.13
dill==0.3.4
docutils==0.18.1
flasgger==0.9.5
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
future==0.18.2
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
humanfriendly==10.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.11.3
importlib-resources==5.6.0
inapppy==2.5.2
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
jsonschema==4.4.0
kombu==5.2.4
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
mccabe==0.7.0
mistune==2.0.2
more-itertools==8.12.0
multidict==6.0.2
oauth2client==4.1.3
packaging==21.3
paho-mqtt==1.6.1
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfunctional==1.4.3
PyJWT==2.3.0
pylint==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.7
pyreadline==2.1
pyrsistent==0.18.1
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
PyYAML==6.0
redis==4.2.0
requests==2.27.1
rope==0.23.0
rsa==4.8
s3transfer==0.5.2
sentry-sdk==1.5.8
six==1.16.0
SQLAlchemy==1.4.32
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
voluptuous==0.12.2
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.0
yarl==1.7.2
zipp==3.7.0
